空間内へ表示するボタンのクラス化
